requirements.txt
flask==3.0.0
openai==1.55.3
python-dotenv==1.0.0
tenacity==8.2.3
requests==2.31.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0